Changed parameter-input from pointers to arrays
